refactor: ContentDto
