+++ src/main/java/com/codeit/sb02mplteam2/domain/content/dto/content/ContentRequestDto.java
@@ -1,9 +1,9 @@
-package com.codeit.sb02mplteam2.domain.content.dto;
+package com.codeit.sb02mplteam2.domain.content.dto.content;
 
 public record ContentRequestDto(
     String title,
     String category,
     String description,
-    Long binaryContentId
+    String imageUrl
 ) {
 }
